backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
folium==0.15.0
geopandas==0.14.1
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
bcrypt==4.0.1
marshmallow==3.20.1
geopy==2.4.0
shapely==2.0.2
